requirements.txt
transformers
mlx
pillow
torch